gitVersion
git-svn-id: file:///E/SVN/yg/trunk@93 062f0958-56b8-ac45-908c-11283a37e8c9

+++ code/tools/fileVersion/Common/CommonBase.cs
@@ -69,8 +69,8 @@ namespace FileVersion
         }
 
         public string folder { get; protected set; }
-        public int highVersion { get; protected set; }
-        public int lowVersion { get; protected set; }
+        public long highVersion { get; protected set; }
+        public long lowVersion { get; protected set; }
 
         public virtual void Run()
         {
